+++ R/plot_classified_syns.R
@@ -40,7 +40,7 @@ plot_classified_syns <- function(x,
                                  dot_size = 0.1,
                                  line_col = "black",
                                  sd_col = "grey80",
-                                 condition,
+                                 condition = NA,
                                  show_plot = TRUE) {
   message("\nCreating synergy plots for condition ", condition, "...")